pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "docxify"
version = "0.1.0"
description = "Convert AI-generated markdown to professional .docx files and back."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "João Rodrigues", email = "jcrodriguesjr@hotmail.com" },
]
keywords = ["markdown", "docx", "ai", "converter", "document", "chatgpt", "claude"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business",
    "Topic :: Text Processing :: Markup :: Markdown",
]
dependencies = [
    "python-docx>=1.0.0",
    "click>=8.0",
    "tomli>=2.0; python_version < '3.11'",
]

[project.scripts]
docxify = "docxify.cli:app"

[project.urls]
Homepage = "https://github.com/jcrodriguesjr0/docxify"
Repository = "https://github.com/jcrodriguesjr0/docxify"
Issues = "https://github.com/jcrodriguesjr0/docxify/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/docxify"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 100
